backend/app/requirements.txt
fastapi==0.115.0
httpx==0.27.2
jose==1.0.0
passlib==1.7.2
pymongo==4.6.0
python_jose==3.3.0
PyYAML==6.0.2
redmail==0.6.0
Requests==2.32.3
pydantic==2.9.2
pydantic[email]
uvicorn
uvicorn[standard]
bcrypt
python-multipart==0.0.5